# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShopFlow Platform API
  description: |
    REST API for the wholesale/retail shop management platform (React web and React Native clients).

    * Authentication: mobile number + OTP → short-lived bearer access token (JWT) plus a rotating refresh token.
    * Authorization: role (OWNER, ADMIN, CUSTOMER) plus permission codes; customers can only access their own data.
    * Envelope: every JSON response is `{ success, data, message, pagination?, error?, requestId }`.
    * Errors: `error.code` is one of the documented business codes (e.g. INSUFFICIENT_STOCK, CREDIT_LIMIT_EXCEEDED).
    * Money: decimal numbers with 2 places, calculated only by the backend.
    * Lists: `page` (1-based), `pageSize` (max 100), `sort=field,asc|desc`.
    * Retries: send `Idempotency-Key` on order creation, payment recording, invoice generation and WhatsApp sends.
  version: v1
servers:
- url: /
  description: Current host
...
paths:
...
  /api/v1/cart/items:
    post:
      tags:
      - Cart
      summary: Add a product
      description: "Adds to the quantity when the product is already in the cart.\
        \ Errors: PRODUCT_INACTIVE."
      operationId: add
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCartResponse"
      security:
      - bearerAuth: []
...
  /api/v1/cart/items/{id}:
    delete:
      tags:
      - Cart
      summary: Remove a line
      operationId: remove_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCartResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Cart
      summary: Change a line quantity
      operationId: update_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCartResponse"
      security:
      - bearerAuth: []
...
  /api/v1/dashboard/customer:
    get:
      tags:
      - Dashboard
      summary: Customer home summary
      description: "Outstanding, open orders, recent orders and recently ordered products\
        \ for the signed-in customer."
      operationId: customer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
      security:
      - bearerAuth: []
...
  /api/v1/cart:
    get:
      tags:
      - Cart
      summary: Get my cart
      description: Lines are re-priced for the customer; checkoutReady=false when
        a line is unavailable or short of stock.
      operationId: get_12
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCartResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Cart
      summary: Empty the cart
      operationId: clear
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldIssue"
...
    FieldIssue:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
    CartItemRequest:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: number
          minimum: 0.001
      required:
      - productId
      - quantity
    ApiResponseCartResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/CartResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    CartLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        sku:
          type: string
        productName:
          type: string
        unit:
          type: string
        imageUrl:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        mrp:
          type: number
        discountAmount:
          type: number
        taxRate:
          type: number
        taxableAmount:
          type: number
        taxAmount:
          type: number
        lineTotal:
          type: number
        stockStatus:
          type: string
        issue:
          type: string
    CartResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartLine"
        itemCount:
          type: integer
          format: int32
        subtotal:
          type: number
        discountTotal:
          type: number
        taxableTotal:
          type: number
        cgstTotal:
          type: number
        sgstTotal:
          type: number
        igstTotal:
          type: number
        roundOff:
          type: number
        grandTotal:
          type: number
        interState:
          type: boolean
        checkoutReady:
          type: boolean
...
    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: number
          minimum: 0.001
      required:
      - quantity
...
    ApiResponseMapStringObject:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: {}
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT